--- src/context/ContextProvider.js
@@ -1,1 +1,0 @@
-const UserProvider = ({ children }) => {};
