+++ src/context/CustomerContext.js
@@ -1,0 +1,5 @@
+import { createContext } from "react";
+
+const CustomerContext = createContext();
+
+export default CustomerContext;
